+++ envd/api/runtime/__init__.py
@@ -84,3 +84,12 @@ def environ(env: Dict[str, str]):
     runtime.environ(env={"ENVD_MODE": "DEV"})
     ```
     """
+
+
+def mount(host_path: str, envd_path: str):
+    """Mount from host path to container path (runtime)
+
+    Args:
+        host_path (str): source path in the host machine
+        envd_path (str): destination path in the envd container
+    """
