+++ examples/dgl/build.envd
@@ -8,8 +8,8 @@ def build():
     # Select the shell environment you like
     shell("zsh")
 
-    # io.mount(host_path="~/.envd/data/dgl", envd_path="~/.dgl")
-    io.mount(host_path=data.envd("dgl"), envd_path=data.path.dgl)
+    # runtime.mount(host_path="~/.envd/data/dgl", envd_path="~/.dgl")
+    runtime.mount(host_path=data.envd("dgl"), envd_path=data.path.dgl)
 
 def build_gpu():
     # Use ubuntu20.04 as base image and install python    
@@ -26,4 +26,4 @@ def build_gpu():
     # Select the shell environment you like
     shell("zsh")
     
-    io.mount(host_path=data.envd("dgl"), envd_path=data.path.dgl)
+    runtime.mount(host_path=data.envd("dgl"), envd_path=data.path.dgl)
