+++ pkg/lang/frontend/starlark/io/const.go
@@ -15,6 +15,5 @@
 package io
 
 const (
-	ruleCopy  = "io.copy"
-	ruleMount = "io.mount"
+	ruleCopy = "io.copy"
 )
