+++ pkg/lang/frontend/starlark/io/io.go
@@ -15,17 +15,10 @@
 package io
 
 import (
-	"os/user"
-	"path/filepath"
-	"strings"
-
-	"github.com/cockroachdb/errors"
-
 	"github.com/sirupsen/logrus"
 	"go.starlark.net/starlark"
 	"go.starlark.net/starlarkstruct"
 
-	"github.com/tensorchord/envd/pkg/lang/frontend/starlark/data"
 	"github.com/tensorchord/envd/pkg/lang/ir"
 )
 
@@ -36,62 +29,8 @@ var (
 var Module = &starlarkstruct.Module{
 	Name: "io",
 	Members: starlark.StringDict{
-		"copy":  starlark.NewBuiltin(ruleCopy, ruleFuncCopy),
-		"mount": starlark.NewBuiltin(ruleMount, ruleFuncMount),
+		"copy": starlark.NewBuiltin(ruleCopy, ruleFuncCopy),
 	},
-}
-
-func ruleFuncMount(thread *starlark.Thread, _ *starlark.Builtin,
-	args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
-	var source starlark.Value
-	var destination starlark.String
-
-	if err := starlark.UnpackArgs(ruleMount, args, kwargs,
-		"host_path?", &source, "envd_path?", &destination); err != nil {
-		return nil, err
-	}
-
-	var sourceStr string
-	var err error
-
-	if v, ok := source.(*data.DataSourceValue); ok {
-		err = v.Init()
-		if err != nil {
-			return starlark.None, err
-		}
-		sourceStr, err = v.GetHostDir()
-		if err != nil {
-			return starlark.None, err
-		}
-	} else if vs, ok := source.(starlark.String); ok {
-		sourceStr = vs.GoString()
-	} else {
-		return starlark.None, errors.New("invalid data source")
-	}
-
-	destinationStr := destination.GoString()
-
-	logger.Debugf("rule `%s` is invoked, src=%s, dest=%s",
-		ruleMount, sourceStr, destinationStr)
-
-	// Expand source directory based on host user
-	usr, _ := user.Current()
-	dir := usr.HomeDir
-	if sourceStr == "~" {
-		sourceStr = dir
-	} else if strings.HasPrefix(sourceStr, "~/") {
-		sourceStr = filepath.Join(dir, sourceStr[2:])
-	}
-	// Expand dest directory based on container user envd
-	dir = "/home/envd/"
-	if destinationStr == "~" {
-		destinationStr = dir
-	} else if strings.HasPrefix(destinationStr, "~/") {
-		destinationStr = filepath.Join(dir, destinationStr[2:])
-	}
-	ir.Mount(sourceStr, destinationStr)
-
-	return starlark.None, nil
 }
 
 func ruleFuncCopy(thread *starlark.Thread, _ *starlark.Builtin,
