+++ pkg/lang/frontend/starlark/runtime/runtime.go
@@ -15,11 +15,16 @@
 package runtime
 
 import (
+	"os/user"
+	"path/filepath"
+	"strings"
+
 	"github.com/cockroachdb/errors"
 	"github.com/sirupsen/logrus"
 	"go.starlark.net/starlark"
 	"go.starlark.net/starlarkstruct"
 
+	"github.com/tensorchord/envd/pkg/lang/frontend/starlark/data"
 	"github.com/tensorchord/envd/pkg/lang/ir"
 )
 
@@ -34,6 +39,7 @@ var Module = &starlarkstruct.Module{
 		"daemon":  starlark.NewBuiltin(ruleDaemon, ruleFuncDaemon),
 		"expose":  starlark.NewBuiltin(ruleExpose, ruleFuncExpose),
 		"environ": starlark.NewBuiltin(ruleEnviron, ruleFuncEnviron),
+		"mount":   starlark.NewBuiltin(ruleMount, ruleFuncMount),
 	},
 }
 
@@ -137,5 +143,58 @@ func ruleFuncEnviron(thread *starlark.Thread, _ *starlark.Builtin,
 
 	logger.Debugf("rule `%s` is invoked, env: %v", ruleEnviron, envMap)
 	ir.RuntimeEnviron(envMap)
+	return starlark.None, nil
+}
+
+func ruleFuncMount(thread *starlark.Thread, _ *starlark.Builtin,
+	args starlark.Tuple, kwargs []starlark.Tuple) (starlark.Value, error) {
+	var source starlark.Value
+	var destination starlark.String
+
+	if err := starlark.UnpackArgs(ruleMount, args, kwargs,
+		"host_path?", &source, "envd_path?", &destination); err != nil {
+		return nil, err
+	}
+
+	var sourceStr string
+	var err error
+
+	if v, ok := source.(*data.DataSourceValue); ok {
+		err = v.Init()
+		if err != nil {
+			return starlark.None, err
+		}
+		sourceStr, err = v.GetHostDir()
+		if err != nil {
+			return starlark.None, err
+		}
+	} else if vs, ok := source.(starlark.String); ok {
+		sourceStr = vs.GoString()
+	} else {
+		return starlark.None, errors.New("invalid data source")
+	}
+
+	destinationStr := destination.GoString()
+
+	logger.Debugf("rule `%s` is invoked, src=%s, dest=%s",
+		ruleMount, sourceStr, destinationStr)
+
+	// Expand source directory based on host user
+	usr, _ := user.Current()
+	dir := usr.HomeDir
+	if sourceStr == "~" {
+		sourceStr = dir
+	} else if strings.HasPrefix(sourceStr, "~/") {
+		sourceStr = filepath.Join(dir, sourceStr[2:])
+	}
+	// Expand dest directory based on container user envd
+	dir = "/home/envd/"
+	if destinationStr == "~" {
+		destinationStr = dir
+	} else if strings.HasPrefix(destinationStr, "~/") {
+		destinationStr = filepath.Join(dir, destinationStr[2:])
+	}
+	ir.Mount(sourceStr, destinationStr)
+
 	return starlark.None, nil
 }
